feat: show map on launch with login icon
